Move feature views out of ui folders
