+++ src/app/canvas/CanvasContainer.tsx
@@ -2,9 +2,9 @@ import React, { useEffect, useState } from 'react'
 import { useNavigationStore } from '@/features/navigation/state'
 import type { View } from '@/shared/types'
 import ChatsView from '@/features/chat/ui/ChatsView'
-import FilesView from '@/features/files/ui/FilesView'
-import ReposView from '@/features/repos/ui/ReposView'
-import BranchesView from '@/features/branches/ui/BranchesView'
+import FilesView from '@/features/files/FilesView'
+import ReposView from '@/features/repos/ReposView'
+import BranchesView from '@/features/branches/BranchesView'
 import HelpView from '@/features/help/HelpView'
 import CustomersView from '@/features/customers/CustomersView'
 import WeatherView from '@/features/weather/WeatherView'
@@ -12,10 +12,10 @@ import HomeView from '@/features/home/HomeView'
 import InnovationsView from '@/features/innovations/InnovationsView'
 import SettingsView from '@/features/settings/SettingsView'
 import AccountView from '@/features/account/AccountView'
-import NappsView from '@/features/napps/ui/NappsView'
+import NappsView from '@/features/napps/NappsView'
 import TranscludesView from '@/features/transcludes/TranscludesView'
 import MessagesView from '@/features/messages/MessagesView'
-import ProcessesView from '@/features/processes/ui/ProcessesView'
+import ProcessesView from '@/features/processes/ProcessesView'
 
 const CanvasContainer: React.FC = () => {
   const currentView = useNavigationStore((state) => state.currentView)
